refactor: useChatHistory custom hook
